endpoint: move code to `identifiers.go`
Signed-off by: Ian Vernon <ian@cilium.io>
