+++ pkg/endpoint/endpoint.go
@@ -999,157 +999,6 @@ func (e *Endpoint) RegenerateWait(reason string) error {
 	return nil
 }
 
-// GetContainerName returns the name of the container for the endpoint.
-func (e *Endpoint) GetContainerName() string {
-	e.unconditionalRLock()
-	defer e.runlock()
-	return e.containerName
-}
-
-// SetContainerName modifies the endpoint's container name
-func (e *Endpoint) SetContainerName(name string) {
-	e.unconditionalLock()
-	e.containerName = name
-	e.unlock()
-}
-
-// GetK8sNamespace returns the name of the pod if the endpoint represents a
-// Kubernetes pod
-func (e *Endpoint) GetK8sNamespace() string {
-	e.unconditionalRLock()
-	ns := e.K8sNamespace
-	e.runlock()
-	return ns
-}
-
-// SetK8sNamespace modifies the endpoint's pod name
-func (e *Endpoint) SetK8sNamespace(name string) {
-	e.unconditionalLock()
-	e.K8sNamespace = name
-	e.UpdateLogger(map[string]interface{}{
-		logfields.K8sPodName: e.getK8sNamespaceAndPodName(),
-	})
-	e.unlock()
-}
-
-// K8sNamespaceAndPodNameIsSet returns true if the pod name is set
-func (e *Endpoint) K8sNamespaceAndPodNameIsSet() bool {
-	e.unconditionalLock()
-	podName := e.getK8sNamespaceAndPodName()
-	e.unlock()
-	return podName != "" && podName != "/"
-}
-
-// GetK8sPodName returns the name of the pod if the endpoint represents a
-// Kubernetes pod
-func (e *Endpoint) GetK8sPodName() string {
-	e.unconditionalRLock()
-	k8sPodName := e.K8sPodName
-	e.runlock()
-
-	return k8sPodName
-}
-
-// HumanStringLocked returns the endpoint's most human readable identifier as string
-func (e *Endpoint) HumanStringLocked() string {
-	if pod := e.getK8sNamespaceAndPodName(); pod != "" {
-		return pod
-	}
-
-	return e.StringID()
-}
-
-// GetK8sNamespaceAndPodName returns the corresponding namespace and pod
-// name for this endpoint.
-func (e *Endpoint) GetK8sNamespaceAndPodName() string {
-	e.unconditionalRLock()
-	defer e.runlock()
-
-	return e.getK8sNamespaceAndPodName()
-}
-
-func (e *Endpoint) getK8sNamespaceAndPodName() string {
-	return e.K8sNamespace + "/" + e.K8sPodName
-}
-
-// SetK8sPodName modifies the endpoint's pod name
-func (e *Endpoint) SetK8sPodName(name string) {
-	e.unconditionalLock()
-	e.K8sPodName = name
-	e.UpdateLogger(map[string]interface{}{
-		logfields.K8sPodName: e.getK8sNamespaceAndPodName(),
-	})
-	e.unlock()
-}
-
-// SetContainerID modifies the endpoint's container ID
-func (e *Endpoint) SetContainerID(id string) {
-	e.unconditionalLock()
-	e.containerID = id
-	e.UpdateLogger(map[string]interface{}{
-		logfields.ContainerID: e.getShortContainerID(),
-	})
-	e.unlock()
-}
-
-// GetContainerID returns the endpoint's container ID
-func (e *Endpoint) GetContainerID() string {
-	e.unconditionalRLock()
-	cID := e.containerID
-	e.runlock()
-	return cID
-}
-
-// GetShortContainerID returns the endpoint's shortened container ID
-func (e *Endpoint) GetShortContainerID() string {
-	e.unconditionalRLock()
-	defer e.runlock()
-
-	return e.getShortContainerID()
-}
-
-func (e *Endpoint) getShortContainerID() string {
-	if e == nil {
-		return ""
-	}
-
-	caplen := 10
-	if len(e.containerID) <= caplen {
-		return e.containerID
-	}
-
-	return e.containerID[:caplen]
-
-}
-
-// SetDockerEndpointID modifies the endpoint's Docker Endpoint ID
-func (e *Endpoint) SetDockerEndpointID(id string) {
-	e.unconditionalLock()
-	e.dockerEndpointID = id
-	e.unlock()
-}
-
-func (e *Endpoint) GetDockerEndpointID() string {
-	e.unconditionalRLock()
-	defer e.runlock()
-	return e.dockerEndpointID
-}
-
-// SetDockerNetworkID modifies the endpoint's Docker Endpoint ID
-func (e *Endpoint) SetDockerNetworkID(id string) {
-	e.unconditionalLock()
-	e.dockerNetworkID = id
-	e.unlock()
-}
-
-// GetDockerNetworkID returns the endpoint's Docker Endpoint ID
-func (e *Endpoint) GetDockerNetworkID() string {
-	e.unconditionalRLock()
-	defer e.runlock()
-
-	return e.dockerNetworkID
-}
-
 // OnProxyPolicyUpdate is a callback used to update the Endpoint's
 // proxyPolicyRevision when the specified revision has been applied in the
 // proxy.
